+++ components/surveymonkey/src/actions/collectors/listCollectors.ts
@@ -12,13 +12,13 @@ export const listCollectors = action({
   inputs: listCollectorsInputs,
   perform: async (
     context,
-    { connection, surveyId, fetchAll, page, perPage },
+    { connection, surveyId, fetchAll, pagination = {} },
   ) => {
     const client = createClient(connection, context.debug.enabled);
     const endpoint = `/surveys/${surveyId}/collectors`;
     const data = await paginateResults<Collector>(client, endpoint, fetchAll, {
-      page,
-      per_page: perPage,
+      page: pagination.page,
+      per_page: pagination.perPage,
     });
     return { data };
   },