+++ components/surveymonkey/src/actions/responses/listResponses.ts
@@ -12,7 +12,7 @@ export const listResponses = action({
   inputs: listResponsesInputs,
   perform: async (
     context,
-    { connection, surveyId, status, fetchAll, page, perPage },
+    { connection, surveyId, status, fetchAll, pagination = {} },
   ) => {
     const client = createClient(connection, context.debug.enabled);
     const endpoint = `/surveys/${surveyId}/responses`;
@@ -26,8 +26,8 @@ export const listResponses = action({
       fetchAll,
       {
         ...params,
-        page,
-        per_page: perPage,
+        page: pagination.page,
+        per_page: pagination.perPage,
       },
     );
     return { data };